# api/api.yml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Review
servers:
  - url: https://review-me.page
    description: Dev server
paths:
  /groups:
    post:
      summary: 리뷰 그룹을 생성합니다.
      operationId: createGroup
      tags:
        - group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreationRequest'
      responses:
        '200':
          description: 리뷰 그룹이 성공적으로 생성되었습니다. 방 ID와 확인코드를 포함합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupCreationResponse'
...
components:
  schemas:
...
    GroupCreationRequest:
      properties:
        revieweeName:
          type: string
          description: 리뷰이 이름
          example: 산초
          minLength: 1
          maxLength: 50
        projectName:
          type: string
          description: 프로젝트 이름
          example: 리뷰미
          minLength: 1
          maxLength: 50
...
    GroupCreationResponse:
      properties:
        reviewRequestCode:
          type: string
          description: 리뷰 요청 코드
          example: VN53pTY
        groupAccessCode:
          type: string
          description: 리뷰 그룹 확인 코드
          example: Pv3n42
...